requirements.txt
Flask==3.1.0
joblib==1.4.2
keras==3.7.0
llama_cpp_python==0.3.5
langdetect==1.0.9
nltk==3.9.1
numpy==2.0.2
opencv_python==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.0.0
pytesseract==0.3.13
scikit_learn==1.6.0
xgboost==2.1.3
